Moved these scripts, now for the rest of the update
