+++ .gitignore
@@ -1,4 +1,3 @@
 aasm
 aasm.exe
 scripts/game_of_life/*
-laasm
