+++ arbnum_global.h
